requirements.txt
uvicorn[standard]
pydantic[email]
fastapi
mysql-connector-python
sqlalchemy
cryptography
ulid-py
python-multipart